+++ views/__init__.py
@@ -1,11 +1,9 @@
-from .post_requests import get_all_posts, get_single_post, create_post, delete_post, create_post_tag, get_certain_post_tags
-from .user_requests import login_user, create_user, get_all_users, login_user
 from .category_requests import create_category, get_all_categories
-from .post_requests import get_all_posts, get_single_post, update_post, get_posts_by_category
-from .tag_requests import create_tag, get_all_tags, get_tags_by_label
-from .user_requests import create_user, get_all_users, login_user, get_single_user
 from .comment_requests import get_all_comments
-
-
-
-
+from .post_requests import (create_post, create_post_tag, delete_post,
+                            get_all_posts, get_posts_by_category,
+                            get_single_post, update_post)
+from .post_tag_requests import get_certain_post_tags
+from .tag_requests import create_tag, get_all_tags, get_tags_by_label
+from .user_requests import (create_user, get_all_users, get_single_user,
+                            login_user)
